+++ tests/acceptance.suite.yml
@@ -8,10 +8,22 @@
 actor: AcceptanceTester
 modules:
     enabled:
+        # wp-browser supplied classes, which provide WordPress specific functions
         - WPDb
         - WPWebDriver
-        - \Helper\Acceptance
+
+        # Codeception supplied classes, which provide assertions and filesystem functions
         - \Codeception\Module\Asserts
+        - \Codeception\Module\Filesystem
+
+        # Our helper classes, which contain functions used across multiple tests.
+        # If you created a new file in tests/_support/Helper/Acceptance, add its namespace and class below,
+        - \Helper\Acceptance\ConvertKitAPI
+        - \Helper\Acceptance\Email
+        - \Helper\Acceptance\GravityForms
+        - \Helper\Acceptance\Plugin
+        - \Helper\Acceptance\ThirdPartyPlugin
+        - \Helper\Acceptance\Xdebug
     config:
         WPDb:
             dsn: '%TEST_SITE_DB_DSN%'
